refactor: rollup config work moved into main config folder
