+++ @remirror/react/package.json
@@ -28,54 +28,7 @@
     "watch:declaration": "tsc -p ./tsconfig.prod.json --emitDeclarationOnly --declarationMap --watch",
     "watch:rollup": "rollup -c -w"
   },
-  "dependencies": {
-    "@types/hoist-non-react-statics": "3.0.1",
-    "@types/lodash": "4.14.120",
-    "@types/prosemirror-commands": "1.0.1",
-    "@types/prosemirror-gapcursor": "1.0.1",
-    "@types/prosemirror-history": "1.0.1",
-    "@types/prosemirror-inputrules": "1.0.2",
-    "@types/prosemirror-keymap": "1.0.1",
-    "@types/prosemirror-menu": "1.0.1",
-    "@types/prosemirror-model": "1.5.1",
-    "@types/prosemirror-schema-basic": "1.0.1",
-    "@types/prosemirror-schema-list": "1.0.1",
-    "@types/prosemirror-state": "1.2.1",
-    "@types/prosemirror-tables": "0.1.3",
-    "@types/prosemirror-transform": "1.1.0",
-    "@types/prosemirror-view": "1.3.0",
-    "@types/react-icons": "2.2.6",
-    "css-in-js-utils": "3.0.0",
-    "csstype": "2.6.2",
-    "hoist-non-react-statics": "3.3.0",
-    "lodash": "4.17.11",
-    "prosemirror-commands": "1.0.7",
-    "prosemirror-dropcursor": "1.1.1",
-    "prosemirror-gapcursor": "1.0.3",
-    "prosemirror-history": "1.0.3",
-    "prosemirror-inputrules": "1.0.1",
-    "prosemirror-keymap": "1.0.1",
-    "prosemirror-menu": "1.0.5",
-    "prosemirror-model": "1.7.0",
-    "prosemirror-schema-basic": "1.0.0",
-    "prosemirror-schema-list": "1.0.2",
-    "prosemirror-state": "1.2.2",
-    "prosemirror-tables": "0.7.10",
-    "prosemirror-transform": "1.1.3",
-    "prosemirror-view": "1.7.0",
-    "simplytyped": "1.7.0"
-  },
-  "peerDependencies": {
-    "@types/react": "^16.7.18",
-    "@types/react-dom": "^16.0.11",
-    "react": "16.8.0-alpha.1",
-    "react-dom": "16.8.0-alpha.1"
-  },
-  "keywords": [
-    "prosemirror",
-    "react",
-    "ssr",
-    "text editor"
-  ],
+  "dependencies": {},
+  "peerDependencies": {},
   "types": "lib/index.d.ts"
 }
